requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
langdetect==1.0.9
Mako==1.3.8
MarkupSafe==3.0.2
mpmath==1.3.0
msgspec==0.18.6
openai==1.57.4
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
